requirements.txt
# FastAPI Stack
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# ML/Data Science
catboost==1.2.2
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2

# Spark
pyspark==3.5.0

# Data Processing
ibis-framework[duckdb]==7.1.0

# Utilities
requests==2.31.0
python-dotenv==1.0.0
